requirements.txt
pyupbit
discord.py
python-dotenv
aiohttp
